extract tests into separate file
