+++ tests/test_glue/test_partition_filter.py
@@ -1,0 +1,263 @@
+import sure  # noqa # pylint: disable=unused-import
+import pytest
+import boto3
+from botocore.client import ClientError
+
+
+from moto import mock_glue
+from . import helpers
+
+
+@mock_glue
+def test_get_partitions_expression_unknown_column():
+    client = boto3.client("glue", region_name="us-east-1")
+    database_name = "myspecialdatabase"
+    table_name = "myfirsttable"
+    values = ["2018-10-01"]
+    columns = [helpers.create_column("date_col", "date")]
+    helpers.create_database(client, database_name)
+
+    helpers.create_table(client, database_name, table_name)
+
+    helpers.create_partition(
+        client, database_name, table_name, values=values, columns=columns
+    )
+
+    with pytest.raises(ClientError) as exc:
+        client.get_partitions(
+            DatabaseName=database_name,
+            TableName=table_name,
+            Expression="unknown_col IS NULL",
+        )
+
+    exc.value.response["Error"]["Code"].should.equal("InvalidInputException")
+    exc.value.response["Error"]["Message"].should.match("Unknown column 'unknown_col'")
+
+
+@mock_glue
+def test_get_partitions_expression_int_column():
+    client = boto3.client("glue", region_name="us-east-1")
+    database_name = "myspecialdatabase"
+    table_name = "myfirsttable"
+    columns = [helpers.create_column("int_col", "int")]
+
+    helpers.create_database(client, database_name)
+
+    args = (client, database_name, table_name)
+    helpers.create_table(*args, partition_keys=columns)
+    helpers.create_partition(*args, values=["1"], columns=columns)
+    helpers.create_partition(*args, values=["2"], columns=columns)
+    helpers.create_partition(*args, values=["3"], columns=columns)
+
+    kwargs = {"DatabaseName": database_name, "TableName": table_name}
+
+    response = client.get_partitions(**kwargs)
+    partitions = response["Partitions"]
+    partitions.should.have.length_of(3)
+
+    int_col_is_two_expressions = (
+        "int_col = 2",
+        "int_col = '2'",
+        "int_col IN (2)",
+        "int_col in (6, '4', 2)",
+        "int_col between 2 AND 2",
+        "int_col > 1 AND int_col < 3",
+        "int_col >= 2 and int_col <> 3",
+        "(int_col) = ((2)) (OR) (((int_col))) = (2)",
+        "int_col IS NOT NULL and int_col = 2",
+        "int_col not IN (1, 3)",
+        "int_col NOT BETWEEN 1 AND 1 and int_col NOT BETWEEN 3 AND 3",
+    )
+
+    for expression in int_col_is_two_expressions:
+        response = client.get_partitions(**kwargs, Expression=expression)
+        partitions = response["Partitions"]
+        partitions.should.have.length_of(1)
+        partition = partitions[0]
+        partition["Values"].should.equal(["2"])
+
+    bad_int_expressions = ("int_col = 'test'", "int_col in (2.5)")
+    for expression in bad_int_expressions:
+        with pytest.raises(ClientError) as exc:
+            client.get_partitions(**kwargs, Expression=expression)
+
+        exc.value.response["Error"]["Code"].should.equal("InvalidInputException")
+        exc.value.response["Error"]["Message"].should.match("is not an integer")
+
+    with pytest.raises(ClientError) as exc:
+        client.get_partitions(**kwargs, Expression="int_col LIKE '2'")
+
+    exc.value.response["Error"]["Code"].should.equal("InvalidInputException")
+    exc.value.response["Error"]["Message"].should.match(
+        "Integral data type doesn't support operation 'LIKE'"
+    )
+
+
+@mock_glue
+def test_get_partitions_expression_decimal_column():
+    client = boto3.client("glue", region_name="us-east-1")
+    database_name = "myspecialdatabase"
+    table_name = "myfirsttable"
+    columns = [helpers.create_column("decimal_col", "decimal")]
+
+    helpers.create_database(client, database_name)
+
+    args = (client, database_name, table_name)
+    helpers.create_table(*args, partition_keys=columns)
+    helpers.create_partition(*args, values=["1.2"], columns=columns)
+    helpers.create_partition(*args, values=["2.6"], columns=columns)
+    helpers.create_partition(*args, values=["3e14"], columns=columns)
+
+    kwargs = {"DatabaseName": database_name, "TableName": table_name}
+
+    response = client.get_partitions(**kwargs)
+    partitions = response["Partitions"]
+    partitions.should.have.length_of(3)
+
+    decimal_col_is_two_point_six_expressions = (
+        "decimal_col = 2.6",
+        "decimal_col = '2.6'",
+        "decimal_col IN (2.6)",
+        "decimal_col in (6, '4', 2.6)",
+        "decimal_col between 2 AND 3e10",
+        "decimal_col > 1.5 AND decimal_col < 3",
+        "decimal_col >= 2 and decimal_col <> '3e14'",
+    )
+
+    for expression in decimal_col_is_two_point_six_expressions:
+        response = client.get_partitions(**kwargs, Expression=expression)
+        partitions = response["Partitions"]
+        partitions.should.have.length_of(1)
+        partition = partitions[0]
+        partition["Values"].should.equal(["2.6"])
+
+    bad_decimal_expressions = ("decimal_col = 'test'",)
+    for expression in bad_decimal_expressions:
+        with pytest.raises(ClientError) as exc:
+            client.get_partitions(**kwargs, Expression=expression)
+
+        exc.value.response["Error"]["Code"].should.equal("InvalidInputException")
+        exc.value.response["Error"]["Message"].should.match("test is not a decimal")
+
+    with pytest.raises(ClientError) as exc:
+        client.get_partitions(**kwargs, Expression="decimal_col LIKE '2'")
+
+    exc.value.response["Error"]["Code"].should.equal("InvalidInputException")
+    exc.value.response["Error"]["Message"].should.match(
+        "Decimal data type doesn't support operation 'LIKE'"
+    )
+
+
+@mock_glue
+def test_get_partitions_expression_string_column():
+    client = boto3.client("glue", region_name="us-east-1")
+    database_name = "myspecialdatabase"
+    table_name = "myfirsttable"
+    columns = [helpers.create_column("string_col", "string")]
+
+    helpers.create_database(client, database_name)
+
+    args = (client, database_name, table_name)
+    helpers.create_table(*args, partition_keys=columns)
+    helpers.create_partition(*args, values=["one"], columns=columns)
+    helpers.create_partition(*args, values=["two"], columns=columns)
+    helpers.create_partition(*args, values=["2"], columns=columns)
+    helpers.create_partition(*args, values=["three"], columns=columns)
+
+    kwargs = {"DatabaseName": database_name, "TableName": table_name}
+
+    response = client.get_partitions(**kwargs)
+    partitions = response["Partitions"]
+    partitions.should.have.length_of(4)
+
+    string_col_is_two_expressions = (
+        "string_col = 'two'",
+        "string_col = 2",
+        "string_col IN (1, 2, 3)",
+        "string_col IN ('1', '2', '3')",
+        "string_col IN ('test', 'two', '3')",
+        "string_col between 'twn' AND 'twp'",
+        "string_col > '1' AND string_col < '3'",
+    )
+
+    string_col_like_two_expressions = (
+        "string_col LIKE 'two'",
+        "string_col LIKE 't_o'",
+        "string_col LIKE 't__'",
+        "string_col LIKE '%wo'",
+        "string_col NOT LIKE '%e' AND string_col not like '_'",
+    )
+
+    for expression in string_col_is_two_expressions:
+        response = client.get_partitions(**kwargs, Expression=expression)
+        partitions = response["Partitions"]
+        partitions.should.have.length_of(1)
+        partition = partitions[0]
+        partition["Values"].should.be.within((["two"], ["2"]))
+
+    for expression in string_col_like_two_expressions:
+        with pytest.warns(match="conversion to regex is experimental"):
+            response = client.get_partitions(**kwargs, Expression=expression)
+            partitions = response["Partitions"]
+            partitions.should.have.length_of(1)
+            partition = partitions[0]
+            partition["Values"].should.equal(["two"])
+
+    with pytest.raises(ClientError) as exc:
+        client.get_partitions(**kwargs, Expression="unknown_col LIKE 'two'")
+
+    exc.value.response["Error"]["Code"].should.equal("InvalidInputException")
+    exc.value.response["Error"]["Message"].should.match("Unknown column 'unknown_col'")
+
+
+@mock_glue
+def test_get_partition_expression_warnings_and_exceptions():
+    client = boto3.client("glue", region_name="us-east-1")
+    database_name = "myspecialdatabase"
+    table_name = "myfirsttable"
+    columns = [
+        helpers.create_column("string_col", "string"),
+        helpers.create_column("int_col", "int"),
+        helpers.create_column("float_col", "float"),
+    ]
+
+    helpers.create_database(client, database_name)
+
+    args = (client, database_name, table_name)
+    helpers.create_table(*args, partition_keys=columns)
+    helpers.create_partition(*args, values=["test", "int", "3.14"], columns=columns)
+
+    kwargs = {"DatabaseName": database_name, "TableName": table_name}
+
+    with pytest.warns(match="Expression filtering is experimental"):
+        response = client.get_partitions(**kwargs, Expression="string_col = 'test'")
+        partitions = response["Partitions"]
+        partitions.should.have.length_of(1)
+        partition = partitions[0]
+        partition["Values"].should.equal(["test", "int", "3.14"])
+
+    with pytest.raises(ClientError) as exc:
+        client.get_partitions(**kwargs, Expression="float_col = 3.14")
+
+    exc.value.response["Error"]["Code"].should.equal("InvalidInputException")
+    exc.value.response["Error"]["Message"].should.match("Unknown type : 'float'")
+
+    with pytest.raises(ClientError) as exc:
+        client.get_partitions(**kwargs, Expression="int_col = 2")
+
+    exc.value.response["Error"]["Code"].should.equal("InvalidStateException")
+    exc.value.response["Error"]["Message"].should.match('"int" is not an integer')
+
+    with pytest.raises(ClientError) as exc:
+        client.get_partitions(**kwargs, Expression="unknown_col = 'test'")
+
+    exc.value.response["Error"]["Code"].should.equal("InvalidInputException")
+    exc.value.response["Error"]["Message"].should.match("Unknown column 'unknown_col'")
+
+    with pytest.raises(ClientError) as exc:
+        client.get_partitions(
+            **kwargs, Expression="string_col IS test' AND not parsable"
+        )
+
+    exc.value.response["Error"]["Code"].should.equal("InvalidInputException")
+    exc.value.response["Error"]["Message"].should.match("Unsupported expression")
